KA-100 Move Item.js to models folder
